Add elm-css to the mix!
This commit adds elm-css as dependency and introduces a new design.  It
was really to add typesafe CSS to this simple project, and reassuring to
(re)discover there's a way to build resilient CSS in Elm: as long as we
triage the types we use in the views, it is easy to make sure the CSS is
in sync.