Dissociate boxes from players
\#3 brings up a good observation: boxes and players were mixed to make
things simple, but at hinsight it is not obvious to me it was a good
idea.  This commit dissociates these.

Also, boxes don't need to know their positioning, only the `Clicked`
event should know what's going on and determine which box has been
clicked based on the index it's been given.